+++ src/main/java/nlp/BasicNLProcessor.java
@@ -59,7 +59,7 @@ public class BasicNLProcessor implements NLProcessor{
     }
 
     @Override
-    public Map<Integer, Candidate> filterByRating(Map<Integer, Candidate> candidates, int confVal) {
+    public Map<Integer, Candidate> filterByRating(Map<Integer, Candidate> candidates, double confVal) {
         return restrictTo(candidates, cand -> cand.getConfidence() >= confVal);
     }
